Fix up bash completion
Seems this has been wrong all along. This fixes up the brew-installed
bash completion, which is really all we actually have (OS X’s built-in
seems to be very lacking, esp. under El Capitan), to be sourced early
on, so that we have it for all other completion scripts.

If we’re not on a brew-powered system, try to source the common path for
bash completion instead.