Adjust p4 settings to .extra to avoid conflicts
Perforce, it seems, considers the P4CLIENT setting to be a
unique value that cannot be used on more than one system.

It corresponds to the name of your workspace, as well.

P4USER and P4PORT are common, and set as such.

As for diff & merge, Kaleidoscope isn’t available on remote
systems, so moving those as well