Update Brewfile to trend towards simply using brew.sh
Mathias dropped using a Brewfile back in mathiasbynens/dotfiles#428, as
it was removed at the time.

It came back as a tap + its own bundle, which I continued using.

However, it’s probably not necessary? And I can simply riff on brew.sh
to simplify merges from upstream.

This gets the current Brewfile into a state where it would be simpler to
drop using it